# a7_openapi.yaml
openapi: 3.0.0
info:
 version: '1.0'
 title: 'Redwood Web API'
 description: 'Web Resources Specification (A7) for Redwood'
servers:
- url: http://lbaw.fe.up.pt
- description: Production server
...
paths:
...
 /admin/users:
   get:
     operationId: R206
     summary: 'R206: View user profiles'
     description: 'Show list of user profiles. Access: ADM'
     tags:
       - 'M02: Individual Profile'

     parameters:
       - in: path
         name: id
         schema:
           type: integer
         required: true

     responses:
       '200':
         description: 'Ok. Show list of user profiles'
...